move redux code into utils
