+++ src/index.js
@@ -1,6 +1,6 @@
 import init from './init'
 import model from './model'
-import { getStore } from './redux/store'
+import { getStore } from './utils/store'
 import { pluginExports } from './core'
 
 export default {
